Actualizacion de secuencias.

+++ facturacion-electronica-servicios/src/main/java/com/facturacion/electronica/repo/IComprobanteVentaRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.facturacion.electronica.modelo.ComprobanteVenta;
 
-public interface IComprobanteVentaRepo extends JpaRepository<ComprobanteVenta, Integer> {
+public interface IComprobanteVentaRepo extends JpaRepository<ComprobanteVenta, Long> {
 
 }
